fix: use longValue() for user id equality check

+++ src/main/java/com/FinFlow/domain/Account.java
@@ -66,7 +66,7 @@ public class Account {
   }
 
   public void checkOwner(Long userId) {
-    if (user.getId() != userId) {
+    if (user.getId().longValue() != userId.longValue()) {
       throw new CustomApiException("계좌 소유자가 아닙니다.");
     }
   }
